# DbConnectCourses/DbConnectCourses/Models/CourseStudent.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
#nullable disable
using System;
using System.Collections.Generic;

namespace DbConnectCourses.Models
{
    public partial class CourseStudent
    {
        public string CourseId { get; set; }
        public string StudentId { get; set; }
        public string Grade { get; set; }

        public virtual Course Course { get; set; }
        public virtual Student Student { get; set; }
    }
}