# DbConnectCourses/DbConnectCourses/Models/Instructor.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
#nullable disable
using System;
using System.Collections.Generic;

namespace DbConnectCourses.Models
{
    public partial class Instructor
    {
        public Instructor()
        {
            Courses = new HashSet<Course>();
        }

        public string InstructorId { get; set; }
        public string InstructorFirstName { get; set; }
        public string InstructorLastName { get; set; }
        public string InstructorPhone { get; set; }
        public string InstructorEmail { get; set; }

        public virtual ICollection<Course> Courses { get; set; }
    }
}