# DbConnectCourses/DbConnectCourses/Models/Student.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
#nullable disable
using System;
using System.Collections.Generic;

namespace DbConnectCourses.Models
{
    public partial class Student
    {
        public Student()
        {
            CourseStudents = new HashSet<CourseStudent>();
            Advisors = new HashSet<Advisor>();
        }

        public string StudentId { get; set; }
        public string StudentFirstName { get; set; }
        public string StudentLastName { get; set; }
        public string StudentEmail { get; set; }
        public string StudentPhone { get; set; }

        public virtual ICollection<CourseStudent> CourseStudents { get; set; }

        public virtual ICollection<Advisor> Advisors { get; set; }
    }
}